added drawable-object.class.js
