+++ js/classes/drawable-object.class.js
@@ -1,0 +1,51 @@
+import { imgCachesObject, areImgCachesReady, loadedCachsArray, canvasHeight } from "../script.js";
+
+export class DrawableObject {
+  img;
+  x = 100;
+  y = 100;
+  width = 100;
+  height = 100;
+  imageCache = {};
+
+  loadImage(path) {
+    this.img = new Image();
+    this.img.src = path;
+  }
+
+  async getImages(pathArray) {
+    pathArray.forEach((path) => {
+      let img = new Image();
+      img.src = path;
+      this.imageCache[path] = img;
+    });
+  }
+
+  draw(ctx) {
+    ctx.drawImage(this.img, this.x, this.y, this.width, this.height);
+  }
+
+  changeCacheStatusToFalse(className) {
+    imgCachesObject[className] = false;
+  }
+
+  changeCachStatusToTrue(className) {
+    imgCachesObject[className] = true;
+  }
+
+  async loadImageCache(arr, className) {
+    this.changeCacheStatusToFalse(className);
+    await this.getImages(arr);
+    this.changeCachStatusToTrue(className);
+  }
+
+  checkImagesCacheLoaded() {
+    let checkInterval = setInterval(() => {
+      if (areImgCachesReady) {
+        this.isImageCacheLoaded = true;
+        clearInterval(checkInterval);
+      }
+    }, 100);
+  }
+}
+
