+++ js/classes/moveable-object.class.js
@@ -1,35 +1,16 @@
 import { imgCachesObject, areImgCachesReady, loadedCachsArray, canvasHeight } from "../script.js";
 
-export class MoveableObject {
-  img;
-  x = 100;
-  y = 100;
-  width = 100;
-  height = 100;
-  imageCache = {};
+import { DrawableObject } from "./drawable-object.class.js"
+
+export class MoveableObject extends DrawableObject {
+  
   otherDirection = false;
   speedY = 2;
   lifeEnergy;
 
   constructor() {
+    super();
     this.lifeEnergy = 100;
-  }
-
-  loadImage(path) {
-    this.img = new Image();
-    this.img.src = path;
-  }
-
-  async getImages(pathArray) {
-    pathArray.forEach((path) => {
-      let img = new Image();
-      img.src = path;
-      this.imageCache[path] = img;
-    });
-  }
-
-  draw(ctx) {
-    ctx.drawImage(this.img, this.x, this.y, this.width, this.height);
   }
 
   drawFrame(ctx) {
@@ -116,28 +97,7 @@ export class MoveableObject {
     ctx.stroke();
   }
 
-  changeCacheStatusToFalse(className) {
-    imgCachesObject[className] = false;
-  }
 
-  changeCachStatusToTrue(className) {
-    imgCachesObject[className] = true;
-  }
-
-  async loadImageCache(arr, className) {
-    this.changeCacheStatusToFalse(className);
-    await this.getImages(arr);
-    this.changeCachStatusToTrue(className);
-  }
-
-  checkImagesCacheLoaded() {
-    let checkInterval = setInterval(() => {
-      if (areImgCachesReady) {
-        this.isImageCacheLoaded = true;
-        clearInterval(checkInterval);
-      }
-    }, 100);
-  }
 
   floatToSurface(item) {
     this.currentMovement = setInterval(() => {
